Using pre-configured tunnel device with CTCI
This patch avoids the use of a setuid program to configure a
tun/tap device on unix.  Instead, the device can be defined and
configured at system boot.

The patch allows for just a single parameter on a CTCI device, the
name of the tunnel device, which must be created and configured
elsewhere, for example by openvpn.

The original two-parameter code is still in place and active.

The devices in the configuration file then look like this:

03e0.2  ctci    tun0

I have this in rc.local to create the tunnel device:

openvpn --mktun --dev tun0 --group staff
ifconfig tun0 192.168.2.65 pointopoint 10.0.0.33
echo 1 >/proc/sys/net/ipv4/ip_forward

I also have this udev rule, but I am not convinced it is still
necessary as permissions are now more relaxed than they were when I
started with Hercules (or openvpn has been fixed in the meantime).

KERNEL=="tun",                  GROUP="staff", MODE="0660"