CTCI confusion with pre-configure and -x option
A CTCI device can be defined with the -x option, which allows the tun
device name to be specifed. Hercifc does the device configuration.
Alternatively, a CTCI device can be defined with a single parameter,
the tun device name. In this case, Hercifc does not do the device
configuration, the device is pre-configured. If the -x option was used
it was wrongly assumed that the device was pre-configured.