The support for Hercules variables causes random segment exceptions, often after the operating system has been started.  Thus the current implementation does not "work".
Disable setenv() on unix when pthreads is active (which it is as soon
as a CPU is defined).  The change sets a mark on the wall over in pttrace.c
and refuses setenv() when that mark is set.  Read comments in hscutl.c
if you disagree.

This should probably also be done on Windows, but I shall leave that
to those who are competent on that platform.

The "fix" is nontrivial.