fix modulesdir/datadir/localedir handling
The current logic tries to expand the path variables at configure time (so that they can be written into config.h) rather than delaying the expansion to make time.

Setup CPPFLAGS with the right -D values so that we can let make expand the path vars and not worry about hacking around it inside of autoconf.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>