fix handling of ltdl/dlopen
Nothing was setting hc_cv_have_dlopen, so initialize that based on the dlopen tests that libtool does for us.

With that in place, we can set hc_cv_have_lt_dlopen based on the included ltdl flag.  This lets people who have a *nix system with libdl avoid the ltdl code entirely.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>