Fix so fthreads/wthreads doesn't require FishIO/HercIO
The Windows threading models OPTION_FTHREADS and OPTION_WTHREADS are completely separate from the OPTION_FISHIO and OPTION_HERCIO I/O models. Neither should require the other. OPTION_FTHREADS may imply OPTION_FISHIO, but it should not require it (and vice versa: use of OPTION_FISHIO should not require OPTION_FTHREADS). You should be able to freely mix and match the two with each other. If that's not true it's a bug that needs fixing.