(Temporary?) Fix to OPTION_NO_INLINE... build option issue:
featall.h current #undef's some OPTION_NO_LINE... vstore/dat build options (so that they are always inlined). Prior to this commit, attempting to #define the options (so the functions are NOT inlined) didn't work. This commit corrects this problem so they may be #define'd or #undef'ed as desired.

Additionally, some new HQA scenarios have been defined related to inlining so we can determine which setting we should choose before removing the build option altogether.