Support preconfigured interfaces with FreeBSD/OS/X
On BSD one opens /dev/tun1.  No ioctl() is necessary.  Code is modified
to select a particular device character file based on the interface number.
That is, the interface tun1 is reached by opening /dev/tun1.  /dev/tun is
a cloning device.  Opening it gets you the next available device.