Avoid a crash in channel.c when device is not fully configured.
Resetting a device with no handler dereferences a null pointer
(channel.c:2042).  A test was added to avoid this.

The root problem, however, is that a subchannel is created
even when the device type is not regognised.  This is
particularly nasty with CTC pairs, because only one half
of the requested subchannels are defined.