Add additional defines to project pre-processor settings:
NOTE: This is purely cosmetic.  Hercules uses makefiles to build with and thus does not use Visual Studio's preprocessor defines.  It only uses the values specifically defined in its makefile.  This change is only so Visual Studio's "Intellisense" feature doesn't visually "disable" (gray out) active code sections in its editor because it thinks it's inactive code, and is thus a purely cosmetic change.