Add FEATURE_S370_CHANNEL to .vcproj pre-processor defines:
NOTE: This is purely cosmetic. Hercules uses makefiles to build with and thus does not use Visual Studio's pre-processor defines. It only uses the values specifically defined in its makefile. This change is only so the Visual Studio 2008 editor doesn't "gray out" (disable) source code because it thinks it's inactive.