Implement Interlocked-access facility 2.
Ivan Warren supplied the fix to change the xIY instructions in esame.c
to be in line with the original 360 instructions so that the storage modification
could be replaced by an atomic update.

The width of the field in the interval timer macros is set to 0 as a single
byte is updated.

Bit 52 of the facilities list is now defined, but it appears that it was
already stored as 1.

Added the test case tests/logicimm.tst and its assemble file.