Reworked fix for interlocked access facility 2 not to support this on windows.
This because MSVC does not support __atomic_and_fetch() et al.

Defined a feature macro for it, and conditionally defined the atomic update
functions to the previous behaviour if the facility macro is undefined (end of
feature.h).

Code in general1, general2, and esame remains unchanged.

Updated the test case to actually issue the long displacement instructions and
have the operands more than 4K away from the base.