Add abs console command to display/alter absolute storage.  Update sske test case accordingly.
This need fixing for r 0 when prefixing is active as the address shown is absolute,
but the prefix is R, which is no longer the case.

The fix by Fish to process r connrectly broke the sske with prefix test cases as
they assumed incorrectly that r displayed absolute sorage.