Fix r command with prefixing.  Add ability to set the prefix register. Add prefix .tst test case.
The pr command is enhanced to support a value as an argument.  This value
is treated the same way the set prefix instruction processes its operand.

The abs and r commands are fixed to process the prefixed addresses correctly
and, for r, to display the real address rather than the absolute.  Some processing
is hoisted out of the loop in dump_real_page.