Suppress msg 90020 for EPERM.  Correct pr msg 2290.  Update tests.
Message 90020 is issued blindly on unixen when not run as root.  I think
there is no point in bothering the user with this, but if you feel otherwise
issuing it once will be sufficient.

The message buffer in pr_cmd (hscemode.c) is increased from 32 to 64 to
accomodate the addressing exception message.  This message really warrants
its own number.

Add address mode tests to SSKE test cases.

Various tweeks to redtest.rexx.