Clean up after Fish's valiant effort.
Coerce ccfixme and ccnowarn to understand the various quirks of
clang (like it defines itself as GCC, but you cannot assume that
a particular GCC level means that CLANG does or does not implement
whatever).  Anyhow, opcode.c no longer drowns us in warnings.

There is still a bit of mopping up after the changes to hscmisc.c.

Add the configuration option --disable-interlocked-access-facility-2
to forcefully suppress the use of any underlying atomic fetch and
update.  Mark and Ivan might like to configure with that until they
get a configuration that implements stdatomic.h correctly.

Change hatomic.h to be sensitive to above.