Add suspend script command.  Silence some warnings.  Make some test cases use suspend.
The suspend command, when semaphore.h is available, uses sem_wait to wait for
sem_post when the last active CPU goes into a disabled wait.

When semaphore.h is not available or the semaphore is not initialised,
suspend behaves exactly like pause.  Thus, you should specify a wait time.