Position for the Linux RUNTEST script command.
This commit contais everything outside script.c.

Since sysblk.scrsem is not set, the code in cpu.c is not executed
inside the if.

Stuff that require Fish's attention is marked FORFISH for easy and
relaxed grep experience.

Note that 37 tests fail on make check.  This is because there is
no RUNTEST command.   For a short while I hope.