Fix git rename failure (sigh!)
This has happened to me before and I don't know what causes it but sometimes a git rename fails to be properly reflected in the repository. Either it's a bug in TortoiseGit or I'm just not doing something properly. I'm not just renaming the file but am instead doing a proper git rename (via TortoiseGit) so I'm not sure what's going on. Anyway, this commit should (hopefully!) fix things.