Add hex dump utility function.
Change file type of SoftFloat-specialise to c so that any change
to this file does not cause a complete recompile.  After all,
function declarations go in .h files, but function definitions go in .c files
(or indeed anything you choose.)

Fix problems with handling NaN for float 32 and float 128, but
tantallizingly not float 64.

Add more conversions to the test case.