__builtin_unreachable fix for gcc/clang:
Code was changed to use 'raise(SIGTRAP)' instead of '__builtin_unreachable()' because:

  a) not all compilers support __builtin_unreachable() whereas virtually all compilers/platforms should support raise(SIGTRAP).

  b) it's important to be notified when code originally believed to be "unreachable" is ever reached!

On Windows the '__assume(0)' compiler builtin causes (level 4) C4702 diagnostic if any code follows the __assume(0) statement (because the __assume(0) statement told the compiler was told to ignore it), but on gcc/clang, __builtin_unreachable() does not any diagnostic warning whatsoever if any code happens to follow the __builtin_unreachable() statement.  Thus we use 'raise(SIGTRAP)' instead.