Let DoConfigure accept configuration options and pass them on.
To do:  Lots.  DoConfigure should be able to obtain SoftFloat if it is
not cloned.  This depends on finding the way Hyperion was cloned (.git/config).

Then if SoftFloat is cloned, but not configured, it should do that too.  (Look for config.h)

Also do autogen if it has not been done.  Need to find the tell-tale.