Update DoConfigure to ensure that SoftFloat-3a is installed and ready before configuring Hyperion.
If it is not installed, SoftFloat is cloned in the same way that Hyperion was.
It is then configured and built using the default options.

Then we we run autogen.sh if there is no 'configure'; generate the object
directory structure (SoftFloat has done some of the groundwork here); and
configure using the arguments specified.