CPU Timer Fix Stopping 2nd Level VM ABENDs
There are (were) problems when running VM 2nd level under VM, in that the 2nd level VM would suffer form ABENDs (EXT105, DSP005 and PRG009).  They would occur intermittently, even on idling systems.  When left that way overnight, the 2nd level VM system would automatically restart, but in the morning the CPDUMP files would show these problems.  Additionally, it was observed that the 2nd level VM system would intermittently go into a CPU loop, exhibiting 100% CPU but on a single CPU only.  The ABENDs were observed to occur at that time.  The analysis of the ABENDs showed a probably incorrect CPU Timer emulation in Hyperion.

This was originally believed to be related to Hyperion issue #217, but this was proven to not be the case.  Rather, the problem has now been determined to be caused by the CPU Timer implementation being based on thread_cputime().  Such CPU Timer emulation relied on three variables which were lacking being treated in a critical section.  Also, basing the CPU Timer emulation on thread_cputime() would imply that a running mainframe CPU Timer would be stopped when the Hyperion host would pre-empt the host CPU thread emulating the mainframe CPU, which is incorrect.  Additionally, the Windows emulation of thread_cputime() suffers from a poor resolution (1/16th of a second).

This commit reverts the CPU Timer emulation to the earlier one, to not rely on thread_cputime(), but on hw_clock() and hw_tod.  The processing around DIAG 204 and CPU MAXRATES has not been altered and still uses thread_cputime_us().