Force libtool to accept .a files to link into shared libs.
On NetBSD, override how libtool decides that a library can be linked into
a shared library. The libSoftFloat.a library is not permissible by the
usual rule, which allows *.so and *_pic.a (because all code in a shared
library must be Position Independent Code).

The result is that the shared libraries are not built and hercules
does not work.

Linux commonly seems to have "pass_all" for this setting.

We want to affect the line deplibs_check_method=... in the copied
libtool script.

Hercules 3 still built a libsoftfloat.so.  Building libSoftFloat as
shared lib again would be the best solution.

In the mean time, let's hope it is at least built as PIC.