Restored "make check" / herctest to working order.
The call to logger_undirect() in hscmisc.c was commented
out, in order for tests to work again, whilst still retaining a
working "sh" command with its output ending up in the Hercules log. However, this needs a better fix, as the Hercules log when shutting down normally now lacks the few last messages in the log (which is less harmful).